Fixed case of error token prepended to EOF
Because an error token is prepended to the list of tokens when a token
match fails (no tokens are consumed when they potentially should be)
I needed to check for an error token before the EOF, in which case the
program is actually valid. It's a nasty workaround that I should
probably look into a more elegant solution for.